fix: drop invalid access test from CI
The test "rt_memory_cannot_be_accessed_in_s" is excluded from
the CI runs, as it breaks now due to the memory region harness
introduced in the SPM [1].

The memory that the NWd can share with an SP needs to be known
and advertised before hand in the SPMC manifest. The test
is sharing memory that belongs to realm to cause a memory
access to the SP. This is was one of the cases that was
meant to be prevented with the referred SPM set of patches.

Excluding the test for now, to unblock the changes. The invalid
access test needs to be redesigned.

[1] https://review.trustedfirmware.org/q/topic:%22ja%252Fmem_region_fix%22+(status:open%20OR%20status:merged)

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I6913ad9003108559a8c29d7841c2a58b5fdf5552