test: normal world can't share root memory
In [1] the invalid access test was removed from the respective
tests xml file, as it was leveraging a security vulnerability in
our SPM implementation.

As part of the same patchstack, we have harnessed and limited the
memory that an SP can use via manifest, or memory sharing operations.
This patch reestablishes the invalid access that was discarded,
this time however only validating that the memory sharing of root
memory fails, which should be indicative that the SP can't access it.

Also set missing dependency to the source file: spm_test_helpers.c
to make file tests-invalid-access.mk.

Future patches will include tests to the GPC.

[1] https://review.trustedfirmware.org/q/topic:%22ja%252Fmem_region_fix%22+(status:open%20OR%20status:merged)

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I6ee21586a2e94810fb4656054b26b2a51c97a544