fix(memory share): reclaim and check memory
If the test relates with FFA_MEM_DONATE do not reclaim or check memory.
The memory is owned by the receiver, and is not accessible to sender.

If operation is lend/share, TFTF should reclaim memory contents first,
and then validate that the SP's access to memory. This is to avoid
fault on lend, as lender doesn't have access to memory until after
the reclaim.

Change-Id: I548b5cf2ac5c2774123f438a02565723470b4d85
Signed-off-by: J-Alves <joao.alves@arm.com>