fix(arm): check for null before calling I/O policy callback
I stumbled across this when passing a non-present image id into
the plat_get_image_source function, which caused it to jump
into nowhere, and fault. The image id was less than the maximum
size of the policy array, but was an empty entry.
This check prevents the jump into nowhere.

Change-Id: If5861f0e09b6b2244aee1ec52ce7c91bb7d62ec8
Signed-off-by: Jimmy Brisson <jimmy.brisson@arm.com>