fix(ff-a): tests for FFA_FEATURES
There are different expectations to the result to the FFA_FEATURES
call, provided the feature IDs for ME, SRI and NPI interrupt from
tftf or cactus.

This patch refactors the test to keep the common part equal.
For the differences, locally define the expected test target
for the FFA_FEATURES call.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I96f18487c80d789cf4e10ceee0591786708bce51