Fix: drivers/gic: Relax GICD_ITARGETSR assertion for unicore
As per the GICv2 arch specification, section 4.3.12 in a uniprocessor
implementation all interrupts target one processor and the
GICD_ITARGETSRs are RAZ/WI. The assert on the gicd_itargets_val in the
gic_v2 driver causes the tftf to halt on uniprocessor systems. Thus,
skip the assert on uniprocessor systems.

Change-Id: I471836b413a7835487de3a5dd9cf0e3b6d28b523
Signed-off-by: Pranjal Shrivastava <praan@google.com>