fix(build): use a grouped target for the tests_list files
By default, when specifying multiple independent targets in a rule, make
will make an equivalent target for each. However, that is not correct
when the recipe makes both targets with the same invocation. In that
case they should be specified as a grouped target so that make knows
that both files will be produced at the same time. This is the case for
the generate_test_list.py script.

This is an attempt to fix a race condition when building with many
cores. Note that the race itself was not reproduced.

Change-Id: I2d976951e39b83dd689a1dda63750679e35f8136
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>