fix(ff-a): reporting support of indirect messaging
Hafnium was not reporting the support of indirect messaging
or direct messaging 2, added in FF-A v1.1 and FF-A v1.2,
respectively, in the returned information of the ABI
FFA_PARTITION_INFO_GET. The patch [1] fixes this issue.

This patch makes equivalent change to the arguments expected
for the FFA_PARTITION_INFO_GET return on cactus partitions.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/31712

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I29ef51a7566a09b6fbeada55725d571f0440fbcd