feat(helpers): introduce renamed and unprefixed sysreg wrapper
Some new instructions are actually accesses to system registers
(typically write-only), although their name or description will not
give this away.
Examples are WFIT/WFET (wait-for-interrupt/exception-with-timeout).

Provide a macro that allows to instantiate such function, but without
the mandatory "write_" prefix that the existing wrappers add.

This allows to define and use those functions even when the toolchain
does not know (yet) about those new instructions.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Change-Id: I78fcc02c594e21998c8695198f811aa76b372cd0