fix(ls64): drop ARMv8.7 compiler requirement
The LD64B and ST64B instructions are part of the ARMv8.7 architecture
extension, and as such are unknown to older toolchains. The LS64 test is
forcing the assembler to use ARMv8.7 extensions when compiling the LS64
tests, which requires a recent toolchain - which the binutils shipping
with Ubuntu 20.04 are not.

Add a macro definition for the ls64b and st64b instructions, and use
that to generation the opcode at build time.
This allows us to drop the ARMv8.7 capable toolchain requirement for
this file.

The generared object is identical before and after this patch.

Change-Id: Ibbf3f9bde26bbfc4d510e7df6f39860208eabf36
Signed-off-by: Andre Przywara <andre.przywara@arm.com>