feat(ls64): improve test
Improve the test coverage a bit, and use more complex bit patterns for
the write-read-back test. This helps to spot corruptions in the values.
Also LS64_ARRAYSIZE does not need to be hardcoded and in a header file,
since it's only used privately by the LS64 test code. Explain how the
value of "8" comes about while at it.

Change-Id: I889e614a4158c5fa5d3b9f5b056d8d10f2681755
Signed-off-by: Andre Przywara <andre.przywara@arm.com>