fix(serror): use custom argument for incrementing elr_elx
Add a custom argument to increment the elr_elx after handling SError.
In some cases, to prevent re-triggering the instruction, ELR needs
to be incremented by 4. In other cases, it may not be necessary.

This argument is passed to the handler, which then decides whether
to increment elr_elx by setting the passed argument accordingly after
handling the SError.

Change-Id: I404f3c5e24f894502a8d00c73649be0b2dd540fa
Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>