fix(lib/pcie): bdf macro
The existing PCIE_CREATE_BDF macro is a non standard way of deriving bdf
value. This fix assigns 3 bits for function number, 5 bits for device
number and 8 bits for bus. This bdf value is used as TDISP function id
while passing it to DSM. Using a wrong bdf value results the TDISP
command to fail.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I31301be4dfc9bd2409da73e54715f73079a921cb