test(fuzz): Fix single feature testing
     This change fixes an issue where a user provides a new feature
that can be fuzzed independently of the current setup of fuzzing.
Currently compile fails will cause issues if single feature fuzzing
is desired and the automated flow will disable sets that are not in
the SMC description file.  There are some small changes to the source
files to enable this behavior for both SDEI and Vendor.  There is also
a change to the automated scripting.

Change-Id: Ic768e9f1b285225f12f23c1e36acb668088ad129
Signed-off-by: Mark Dykes <mark.dykes@arm.com>
Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>