fix(smccc): availability test: add two features and fix TRNDR
The SMCCC_ARCH_FEATURE_AVAILABILITY test was not up-to-date and was
missing two features: FEAT_TWED and FEAT_PMUV3P9. Connect the SCR_EL3
and MDCR_EL3 bits to their corresponding ID register fields, so that
they can be tested.

At the same time the FEAT_RNG_TRAP test was slightly off: the SMCCC spec
says it should report accessibility of the RNDR and RNDRRS registers, so
we should look at FEAT_RNG, not FEAT_RNG_TRAP when checking the TRNDR
bit.

This fixes the tf-a-tests run on an FVP with ARMv9.4 enabled, which was
reporting the following issues before:
=====================================
> Executing 'SMCCC_ARCH_FEATURE_AVAILABILITY test'
  TEST COMPLETE                                                 Failed
is_feat_rng_trap_present says feature is supported but SCR_TRNDR_BIT was
	not set!
SCR_EL3 still has values set: 0x20000000. Test needs to be updated
MDCR_EL3 still has values set: 0x80. Test needs to be updated
=====================================

Change-Id: I73a0d240b2cd1a16e1c64d3d66ee30e658c9c946
Signed-off-by: Andre Przywara <andre.przywara@arm.com>