feat(plat/rdaspen): introduce rdaspen platform support
Add support in TFTF for the CSS automotive platform RD-Aspen.

RD-Aspen uses the GICv3 driver and implements a 4-cluster topology
with 4 cores per cluster, for a total of 16 cores.

For platform details, see the TF-A RD-Aspen documentation:
https://trustedfirmware-a.readthedocs.io/en/latest/plat/arm/automotive_rd/rdaspen.html

Creates the skip list for RD-Aspen, since it uses OP-TEE and the
FF-A Interrupt and FF-A Notifications tests depend on Cactus
managed-exit behavior, which is not available in this setup.

Change-Id: I363289181f40967496d6bd85b8dd40f40871cad0
Signed-off-by: Ahmed Azeem <ahmed.azeem@arm.com>
Signed-off-by: Sanjana Virupakshagouda <sanjana.virupakshagouda@arm.com>
Signed-off-by: Ahmed Tiba <ahmed.tiba@arm.com>