feat(rme): update FEAT_MPAM tests on Realms
Currently, to test that accessing a FEAT_MPAM register from a Realm
causes an undefined abort injected back to the Realm, we only test
by accessing a single register.

This patches updates the test by trying to access all MPAM registers
from the Realm to validate that an undefined abort is taken to the
Realm for all the registers.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I16c88d467eb2a49342694536a1c7b6358416dc34