fix(versal2): update test skip list
Skipping tests related to invalid entry addresses and system suspend
with cores suspended. This will continue until the implementation in
TF-A is completed to properly handle invalid entry addresses and
resolve system suspend failures.
Additionally, skipping tests for FFA Group0 in the normal world since
FFA support is not available.

Change-Id: Icc62f563ffb0321f7cdc2e7399a3c3b6c908d778
Signed-off-by: Maheedhar Bollapalli <maheedharsai.bollapalli@amd.com>