feat(rdaspen/ras): add helpers and tests for cpu ras
This introduces RAS helpers from the specification for injecting
CPU RAS correctable errors for Cortex-A720AE IP from EL2 NS, and
adds RD-Aspen specific test cases for CPU RAS CE, DE, and UC.

The tests inject single-bit correctable, deferred, and
uncorrectable RAS faults using the pseudo-fault and misc
registers.

The test is platform-specific and must be built with
TESTS=rdaspen, which builds the CE and DE cases. If the user
wishes to inject an uncontainable UC error, it can be built with
TESTS=rdaspen-uncontainable.

The uncontainable test is separated because it can only be tested
alone, since context does not return from EL3 once injected.

These tests do not run as part of the common test framework.

Change-Id: I70d2f2752156c121087558e9ba19e59233272338
Signed-off-by: Ahmed Azeem <ahmed.azeem@arm.com>
Signed-off-by: Ahmed Tiba <ahmed.tiba@arm.com>