fix(realm): add missing size definitions for Juno
Fixes compilation failure for Juno targets by ensuring the necessary
size definitions from `common_def.h` are included. This resolves build
errors even when the tests are not executed. These tests dynamically
check whether RME is enabled on a platform and run conditionally. As
part of the standard test suite, the source files cannot be excluded
from the TFTF build due to their automatic generation in the test list.
This change ensures that the tests compile successfully for Juno.

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I001a89108c6962a37b24bd3f17e68010a161aaca