feat(amd): add platform specific test cases
Add test cases in the TF-A test framework to validate the
functionality of the EEMI APIs.
Also add client infrastructure to call PM APIs.

-get_api_version
-get_chipid
-feature_check

Note: As of now these changes are just for versal platform. We'll
refactor some of files in future to make it generic for all the
platforms.

Change-Id: Id55de9eadaeaf96ccab4aaa1a494620108c25a19
Signed-off-by: Madhav Bhatt <madhav.bhatt@amd.com>