feat(amd): add test for IOCTL EEMI API
Add a new test case for the IOCTL EEMI API within the TF-A test
framework. The objective of this test is to verify the functionality
and behavior of the IOCTL commands.

Test cover following IOCTL IDs:
- IOCTL_GET_RPU_OPER_MODE
        - To get the RPU operating mode like split or lockstep
- IOCTL_SET_RPU_OPER_MODE
        - To set the RPU operating mode like split or lockstep

These IOCTL Ids require RPU to be enabled either as the default
subsystem or as a RPU subsytem.

Note: This initial test does not cover all IOCTL IDs, as the purpose
of the test is to validate the IOCTL EEMI API interface.

Change-Id: I9c95facaab171af862ad633f7fdcab0b18b7f9e4
Signed-off-by: Madhav Bhatt <madhav.bhatt@amd.com>