feat(amd): add test for pin EEMI APIs
Add test for the pin-related EEMI APIs within the TF-A test
framework. The purpose of these tests is to validate the
functionality and reliability of various pin operations.

Test cover the following EEMI APIs:
- xpm_pinctrl_request
- xpm_pinctrl_release
- xpm_pinctrl_set_function
- xpm_pinctrl_get_function
- xpm_pinctrl_set_parameter
- xpm_pinctrl_get_parameter

Note: This initial test does not cover all the pin ids, as the
purpose of the test is to validate the pin EEMI APIs interface.

Change-Id: I50142ee110e5cb7427ffaeae7f413aa0e6662006
Signed-off-by: Madhav Bhatt <madhav.bhatt@amd.com>