feat(amd): add test for system shutdown EEMI API
Add test for xpm_system_shutdown EEMI API within the TF-A test
framework. The purpose of this test is for a subsystem to
shutdown self or restart self, Ps or system.

Note: Only the subsystem restart scenario is executed in this test.
Performing a system shutdown would prevent the execution of
any subsequent tests.

Change-Id: Iaf8a7cad96ad6b6137187214c902434f90a78802
Signed-off-by: Madhav Bhatt <madhav.bhatt@amd.com>