fix(spe): turn assert into warning for newer SPE versions
FEAT_SPE gets some updates in the architecture from time to time,
increasing the value in the ID_AA64DFR0_EL1.PMSVer field. The Arm ARM
promises that higher numbers only add features, so any tests for a lower
version should always be valid, even if the hardware supports more
features.

Turn the assert into a warning, so we still log that the test might need
to be upgraded, but don't fail the test anymore.

Change-Id: I89a45bec9b22dae15f754a22f310fc6ea8def714
Signed-off-by: Andre Przywara <andre.przywara@arm.com>