refactor: check for partition info descriptor size and count
This patch extends the "ffa_partition_info_regs_helper" helper to check
for the size and total count of the descriptor count returned by the
invocation of FFA_PARTITION_INFO_GET_REGS ABI.

This is particularly important because the enhancements made in FF-A
v1.3 specification significantly change the descriptors returned by the
above ABI.

Currently, the descriptors follow the format described in FF-A v1.2
spec. In subsequent patch, we upgrade the helper utilities to be
comptatible with the extended descriptors defined in v1.3 spec.

Change-Id: Ib31b8e17f19ac362e7300347ad6a60d0ba47964a
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>