fix(errata_abi): update EM_HIGHER_EL_MITIGATION to avoid failure
 The EM-ABI test for Versal Net was failing on Cortex-A78 AE erratum
2376748. TF-A reports EM_HIGHER_EL_MITIGATION (ret0 == 3) for this
 errata, but tftf logic only passed NOT_AFFECTED outside
the revision range and for higher el mitigation as failure.

Updated errata id 2376748 rxpx_high to 0x02.

EM-ABI test for ZynqMP was failing on Cortex-A53 erratum 835769
and 843419. TF-A reports EM_NOT_AFFECTED for both the errata id's
while rxpx is inside the range resulting in tftf marking tests as
failure.

Updated errata ids 835769 and 843419 rxpx_high to 0x03.

Change-Id: I5191a8683fafe69170d7c49d724c1c6a32d88f59
Signed-off-by: Maheedhar Bollapalli <maheedharsai.bollapalli@amd.com>