feat(amu): make group 1 RAZ testing generic
The AMCR_EL0.CG1RZ bit is a hardware feature that will make sure that
group 1 counters read as zero from lower ELs. This is the method
firmware should use to prevent reads of these counters. So instead of
doing an IMP DEF magic dance to increment the counters, check this bit
and rely on hardware to do the rest.

Change-Id: Ib1997fe17f4e6a34337153bb2ece07ccbaec40cf
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>