fix(fvp): make the FEAT_LS64 test easier to run
The fvp doesn't have any regular memory that supports FEAT_LS64
instructions. Instead it has this weird FIFO device that is the only
thing that can be targeted with an LS64 instruction. Depending on the
bp.ls64_testing_fifo.op_type argument, this device will either return
the result or return it NOTed. Quite annoyingly, the default is to
return with NOTing.

So check both the original value and its bitwise negative before failing
the test. This will allow the test to run regardless of the value of
bp.ls64_testing_fifo.op_type.

Change-Id: Ifd26f2a735e96ce4953e3d1cb9306c642750eca2
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>