refactor: Change undef injection test to FEAT_GCS
Currently the undef injection test relies on ENABLE_FEAT_FGT=0 to enable
trapping of the HFGITR_EL2 register and test the injection mechanism.
However, this cannot be used to test RF-A since it always assumes a
newer architecture version where FEAT_FGT is present and never trapped.
This commit refactors the test to rely on FEAT_GCS and try reading the
GCSCR_EL2 register instead, which can be trapped both in TF-A and RF-A.

Change-Id: I3c53cc36f49d6c6b2ce10be4041904aa35f46042
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>